+++ src/VBox/Runtime/r3/win/errvars-win.cpp
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2011-2015 Oracle Corporation
+ * Copyright (C) 2011-2016 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -28,7 +28,7 @@
 /*********************************************************************************************************************************
 *   Header Files                                                                                                                 *
 *********************************************************************************************************************************/
-#include <Winsock2.h>
+#include <iprt/win/winsock2.h>
 #include <errno.h>
 
 #include <iprt/err.h>
@@ -76,7 +76,7 @@ RTDECL(bool) RTErrVarsHaveChanged(PCRTERRVARS pVars)
     Assert(pVars->ai32Vars[0] == RTERRVARS_MAGIC);
 
     return pVars->ai32Vars[0] != RTERRVARS_MAGIC
-        || pVars->ai32Vars[1] != GetLastError()
+        || (uint32_t)pVars->ai32Vars[1] != GetLastError()
         || pVars->ai32Vars[2] != WSAGetLastError()
         || pVars->ai32Vars[3] != errno;
 }
